requirements.txt
affine==2.4.0
asgiref==3.8.1
attrs==25.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.9
Django==5.0.3
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.1
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.4.0
filelock==3.16.1
fiona==1.10.1
geopandas==1.0.1
gunicorn==23.0.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pipenv==2024.0.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.8.0
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
rasterio==1.4.3
rasterstats==0.20.0
requests==2.32.3
result==0.16.1
returns==0.22.0
setuptools==75.8.0
shapely==2.0.7
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.28.1
whitenoise==6.7.0
